Europeana curator: do not check if an image is available
This removes a check where the curator driver would make an HTTP head request to see if an image in Europeana was online or not as it was common that an image would be broken. Since this implementation, Europeana has implemented a link policy and tooling that removes data with broken links within three months causing the number of broken image URLs to drop.

https://pro.europeana.eu/post/addressing-broken-links-to-digital-objects-our-approach

While there might still potentially be cases of broken links, removing this check gives a major speed boast to the Europeana search.

Side-affect: removed an unused function "europeana_find_photo_by_url()"